Shorten custom environment name; update build script and base.
The name of the user is prepended to the environment name by default, so we can
shorten it here. The base of our build is the ubuntu2004 image. Our curriefense
project is all image based, so there isn't anything to build in advance. The script
isn't really a bash script as much as a list of commands. We can omit the #!.